backend/requirements.txt
typer
python-dotenv
elevenlabs
google-generativeai
fastapi
uvicorn
python-multipart
requests
soundfile
numpy
